tagged items. removed 'AF' prop from ammunitions. other manual fixes.
tagged a BUNCH of 1-to-2 word item references in enhancedItem's converted entry texts after a converter upgrade. manually fixed several small things (regressions from that conversion). removed all of the 'AF' itemProperties from ammunition items (AF means an item USES ammo, not IS ammo... need to fix that with a conditional in the converter).