adds live text, adds alternatives checker
adds live test displayed while the user is speaking, adds a setting to toggle live text mode on/off, and adds logic to pre-parse the interim transcript before displaying it, to try to more closely show what the final interpretation will be. ex: instead of showing Goblin 3/4, it'll try to guess and show Goblin 3 4. ALSO adds logic to test speech recognition alternatives for probable correct interpretations... ex: Goblin tutu had the highest confidence, but another alternative interpretted as Goblin 2 2 so it was probably that one.